+++ lib/widgets/workout_drawer.dart
@@ -7,7 +7,9 @@ import 'package:wiseworkout/core/extensions/context_extension.dart';
 import 'package:wiseworkout/core/extensions/string_extension.dart';
 import 'package:wiseworkout/core/themes/constantes.dart';
 import 'package:wiseworkout/features/di/services/service_locator.dart';
+import 'package:wiseworkout/features/history/screens/monthly_summary_screen.dart';
 import 'package:wiseworkout/features/more/screens/about_us_screen.dart';
+import 'package:wiseworkout/features/more/screens/maintenance_screen.dart';
 import 'package:wiseworkout/features/settings/enums/enums.dart';
 import 'package:wiseworkout/features/settings/signals/locale_store.dart';
 import 'package:wiseworkout/features/settings/signals/theme_mode_store.dart';
@@ -169,9 +171,12 @@ class WorkoutDrawer extends StatelessWidget {
           ListTile(
             leading: Icon(Icons.query_stats, color: Theme.of(context).colorScheme.primary),
             title: Text(context.localizations.drawerMonthlySummaryText),
-            onTap: () {
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(content: Text(context.localizations.drawerMonthlySummaryText)),
+            onTap: () async {
+              await Navigator.push(
+                context,
+                MaterialPageRoute<void>(
+                  builder: (_) => const MonthlySummaryScreen(),
+                ),
               );
             },
           ),
@@ -180,9 +185,12 @@ class WorkoutDrawer extends StatelessWidget {
           ListTile(
             leading: Icon(Icons.home_repair_service, color: Theme.of(context).colorScheme.primary),
             title: Text(context.localizations.titleMaintenanceScreen),
-            onTap: () {
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(content: Text(context.localizations.titleMaintenanceScreen)),
+            onTap: () async {
+              await Navigator.push(
+                context,
+                MaterialPageRoute<void>(
+                  builder: (_) => const MaintenanceScreen(),
+                ),
               );
             },
           ),